Split sessions and connect into separate folder to be accessed via an index.js file
- Rename DB_TYPE to DB_TYPES in constants.js
- Change EMPTY to NONE
