+++ server/config/appConfig.js
@@ -1,13 +1,13 @@
-var DB_TYPE = require('./constants').DB_TYPE;
+var DB_TYPES = require('./constants').DB_TYPES;
 
 /*
  * Set DB_TYPE to a database of your choice:
  * - MONGO: MongoDB
  * - POSTGRES: Postgresql
- * - EMPTY: There is no DB
+ * - NONE: There is no DB connection
  */
 module.exports = {
-  DB_TYPE: process.env.DB_TYPE || DB_TYPE.MONGO,
+  DB_TYPE: process.env.DB_TYPE || DB_TYPES.MONGO,
   ENV: process.env.NODE_ENV
 };
