+++ server/config/connect/index.js
@@ -1,0 +1,7 @@
+var DB_TYPES = require('../constants').DB_TYPES;
+
+module.exports = {
+  [DB_TYPES.MONGO]: require('./mongo'),
+  [DB_TYPES.NONE]: function(){}
+};
+
